+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `#35`: As scp user, I want to choose skip download if local file has bigger size. (`option.skipIfBigger` default to `false`)
 - `#36`: As scp user, I want to choose skip download if local file has smaller size. (`option.skipIfSmaller` default to `false`)
 - `#85`: bug: As scp user, I want to remove mutual exclusive check on file size filters.
+- `#87`: bug: As scp user, I want to remove mutual exclusive check on file modified time filters.
 
 ## `0.8.0`